reorganize the file structure
1. inside the src/helpers directory, it contains:
   -- columnlib directory:
         -- header : contains all column template files
        -- cell: cell type
        -- footer: footer type
        -- modal: column type related popup modal
   -- section directory:
        -- header : related section header component
        -- modal: section popup modal
  -- row directory:
         cell type related files, it  may overlap with the column type/cell directory

2. I moved related files into the above directory

3. in future development, we should each column type has a set of seperate files (e.g., inside subdirectorties under the columnlib directory); aovid to compose several components into one js file.
