+++ src/helpers/row/ExampleImage.js
